Slight refactor. Remove extra whitespace. Make sure path exisits before proceeding

+++ Utility/Resource.py
@@ -491,19 +491,20 @@ def installSslCerts():
         try:
             STAT_0o775 = ( stat.S_IRUSR | stat.S_IWUSR | stat.S_IXUSR
                 | stat.S_IRGRP | stat.S_IWGRP | stat.S_IXGRP
-                | stat.S_IROTH |                stat.S_IXOTH )
+                | stat.S_IROTH | stat.S_IXOTH )
             openssl_dir, openssl_cafile = os.path.split(
                 ssl.get_default_verify_paths().openssl_cafile)
-            import certifi
-            # change working directory to the default SSL directory
-            os.chdir(openssl_dir)
-            relpath_to_certifi_cafile = os.path.relpath(certifi.where())
-            try:
-                os.remove(openssl_cafile)
-            except FileNotFoundError:
-                pass
-            os.symlink(relpath_to_certifi_cafile, openssl_cafile)
-            os.chmod(openssl_cafile, STAT_0o775)
+            if os.path.exists(openssl_dir):
+                import certifi
+                # change working directory to the default SSL directory
+                os.chdir(openssl_dir)
+                relpath_to_certifi_cafile = os.path.relpath(certifi.where())
+                try:
+                    os.remove(openssl_cafile)
+                except FileNotFoundError:
+                    pass
+                os.symlink(relpath_to_certifi_cafile, openssl_cafile)
+                os.chmod(openssl_cafile, STAT_0o775)
         except Exception as e:
             ApiToolLog().LogError(e)
         os.chdir(base_path)
